Make a pass at mmm.h
