+++ src/hatrack/array/flexarray.c
@@ -26,6 +26,7 @@
 #include "hatrack/flexarray.h"
 #include "hatrack/malloc.h"
 #include "hatrack/mmm.h"
+#include "hatrack/hatomic.h"
 #include "hatrack/hatrack_common.h"
 #include "../hatrack-internal.h"
