+++ src/hatrack/array/vector.c
@@ -23,6 +23,7 @@
 #include "hatrack/debug.h"
 #include "hatrack/malloc.h"
 #include "hatrack/mmm.h"
+#include "hatrack/hatomic.h"
 #include "hatrack/hatrack_common.h"
 
 #define VECTOR_MIN_STORE_SZ_LOG 4
