+++ src/hatrack/hash/ballcap.c
@@ -47,6 +47,8 @@
  */
 
 #include "hatrack/ballcap.h"
+#include "hatrack/malloc.h"
+#include "hatrack/hatomic.h"
 #include "../hatrack-internal.h"
 
 #include <stdlib.h>
