+++ src/hatrack/hash/crown.c
@@ -161,6 +161,8 @@
  */
 
 #include "crown-internal.h"
+#include "hatrack/malloc.h"
+#include "hatrack/hatomic.h"
 #include "../hatrack-internal.h"
 
 #include <stdlib.h>
