+++ src/hatrack/hash/hihat-a.c
@@ -37,6 +37,8 @@
  */
 
 #include "hihat-internal.h"
+#include "hatrack/malloc.h"
+#include "hatrack/hatomic.h"
 #include "../hatrack-internal.h"
 
 #include <stdlib.h>
