+++ src/hatrack/hash/hihat.c
@@ -25,6 +25,8 @@
  */
 
 #include "hihat-internal.h"
+#include "hatrack/malloc.h"
+#include "hatrack/hatomic.h"
 #include "../hatrack-internal.h"
 
 #include <stdlib.h>
