+++ src/hatrack/hash/lohat-a.c
@@ -25,6 +25,8 @@
  */
 
 #include "hatrack/lohat-a.h"
+#include "hatrack/malloc.h"
+#include "hatrack/hatomic.h"
 #include "../hatrack-internal.h"
 
 #include <stdlib.h>
