+++ src/hatrack/hash/lohat.c
@@ -26,6 +26,8 @@
  */
 
 #include "hatrack/lohat.h"
+#include "hatrack/malloc.h"
+#include "hatrack/hatomic.h"
 #include "../hatrack-internal.h"
 
 #include <stdlib.h>
