+++ src/hatrack/hash/newshat.c
@@ -25,6 +25,8 @@
  */
 
 #include "hatrack/newshat.h"
+#include "hatrack/malloc.h"
+#include "hatrack/hatomic.h"
 #include "../hatrack-internal.h"
 
 #include <stdlib.h>
