+++ src/hatrack/hash/oldhat.c
@@ -65,6 +65,8 @@
  */
 
 #include "hatrack/oldhat.h"
+#include "hatrack/malloc.h"
+#include "hatrack/hatomic.h"
 #include "../hatrack-internal.h"
 
 #include <stdlib.h>
