+++ src/hatrack/hash/refhat.c
@@ -21,6 +21,8 @@
  */
 
 #include "hatrack/refhat.h"
+#include "hatrack/malloc.h"
+#include "hatrack/hatomic.h"
 #include "../hatrack-internal.h"
 
 #include <stdlib.h>
