+++ src/hatrack/hash/swimcap.c
@@ -34,6 +34,8 @@
  */
 
 #include "hatrack/swimcap.h"
+#include "hatrack/malloc.h"
+#include "hatrack/hatomic.h"
 #include "../hatrack-internal.h"
 
 #include <stdlib.h>
