+++ src/hatrack/hash/tiara.c
@@ -37,6 +37,8 @@
  */
 
 #include "hatrack/tiara.h"
+#include "hatrack/malloc.h"
+#include "hatrack/hatomic.h"
 #include "../hatrack-internal.h"
 
 #include <stdlib.h>
