+++ src/hatrack/hash/witchhat.c
@@ -33,6 +33,8 @@
  */
 
 #include "witchhat-internal.h"
+#include "hatrack/malloc.h"
+#include "hatrack/hatomic.h"
 #include "../hatrack-internal.h"
 
 #include <stdlib.h>
