+++ src/hatrack/hash/woolhat.c
@@ -27,6 +27,8 @@
  */
 
 #include "hatrack/woolhat.h"
+#include "hatrack/malloc.h"
+#include "hatrack/hatomic.h"
 #include "../hatrack-internal.h"
 
 #include <stdlib.h>
@@ -62,13 +64,13 @@ static inline bool      woolhat_need_to_help (woolhat_t *);
 static uint64_t         woolhat_set_ordering (woolhat_record_t *, bool);
 static inline void      woolhat_new_insertion(woolhat_record_t *);
 
+// clang-format on
 
 void
 hatrack_set_view_delete(hatrack_set_view_t *view, uint64_t num)
 {
     hatrack_free(view, sizeof(hatrack_set_view_t) * num);
 }
-
 
 static uint64_t
 woolhat_set_ordering(woolhat_record_t *record, bool deleted_below)
@@ -78,12 +80,12 @@ woolhat_set_ordering(woolhat_record_t *record, bool deleted_below)
     mmm_hdr = mmm_get_header(record);
 
     if (mmm_hdr->create_epoch) {
-	return mmm_hdr->create_epoch;
+        return mmm_hdr->create_epoch;
     }
 
     if ((!record->next) || deleted_below) {
-	mmm_hdr->create_epoch = mmm_hdr->write_epoch;
-	return mmm_hdr->create_epoch;
+        mmm_hdr->create_epoch = mmm_hdr->write_epoch;
+        return mmm_hdr->create_epoch;
     }
 
     mmm_hdr->create_epoch = woolhat_set_ordering(record->next, false);
@@ -102,8 +104,6 @@ woolhat_new_insertion(woolhat_record_t *record)
 
     return;
 }
-
-// clang-format on
 
 woolhat_t *
 woolhat_new(void)
