+++ src/hatrack/queue/capq.c
@@ -48,6 +48,7 @@
 #include "hatrack/capq.h"
 #include "hatrack/malloc.h"
 #include "hatrack/mmm.h"
+#include "hatrack/hatomic.h"
 #include "hatrack/hatrack_common.h"
 #include "../hatrack-internal.h"
