+++ src/hatrack/queue/hatring.c
@@ -23,6 +23,7 @@
 #include "hatrack/hatring.h"
 #include "hatrack/malloc.h"
 #include "hatrack/mmm.h"
+#include "hatrack/hatomic.h"
 #include "hatrack/hatrack_common.h"
 
 #include <string.h>
