+++ src/hatrack/queue/hq.c
@@ -116,6 +116,7 @@
 #include "hatrack/hq.h"
 #include "hatrack/malloc.h"
 #include "hatrack/mmm.h"
+#include "hatrack/hatomic.h"
 #include "hatrack/hatrack_common.h"
 #include "../hatrack-internal.h"
