+++ src/hatrack/queue/logring.c
@@ -24,6 +24,7 @@
 #include "hatrack/logring.h"
 #include "hatrack/malloc.h"
 #include "hatrack/mmm.h"
+#include "hatrack/hatomic.h"
 #include "hatrack/hatrack_common.h"
 
 #include <string.h>
