+++ src/hatrack/queue/q64.c
@@ -23,6 +23,7 @@
 #include "hatrack/q64.h"
 #include "hatrack/malloc.h"
 #include "hatrack/mmm.h"
+#include "hatrack/hatomic.h"
 #include "hatrack/hatrack_common.h"
 #include "../hatrack-internal.h"
