+++ src/hatrack/queue/queue.c
@@ -22,6 +22,7 @@
 #include "hatrack/queue.h"
 #include "hatrack/malloc.h"
 #include "hatrack/mmm.h"
+#include "hatrack/hatomic.h"
 #include "hatrack/hatrack_common.h"
 #include "../hatrack-internal.h"
