+++ src/hatrack/queue/stack.c
@@ -85,6 +85,7 @@
 #include "hatrack/stack.h"
 #include "hatrack/malloc.h"
 #include "hatrack/mmm.h"
+#include "hatrack/hatomic.h"
 #include "hatrack/hatrack_common.h"
 #include "../hatrack-internal.h"
